Xcode project: Set 'Skip install' flag on all targets
If you try to Build & Archive on a Mac or iOS project (which is how you
submit to the AppStores) with Bento4 as a subproject, Bento4 will get
entangled in that archive, making it un-submittable. Unless you
use the Build & Archive feature for Bento4, I suggest disabling archiving
for this library to ease integration into Mac and IOS projects.